+++ lab9/andytsai2000/ans/ans6.txt
