+++ lab9/andytsai2000/cw/account.go
@@ -1,0 +1,12 @@
+package cw
+
+type Account struct {
+}
+
+type PTTAccount struct {
+	Account
+}
+
+type FBAccount struct {
+	Account
+}
