+++ lab9/andytsai2000/cw/system.go
@@ -1,0 +1,45 @@
+package cw
+
+import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+)
+
+func checkErr(e error) {
+	if e != nil {
+		fmt.Println(e.Error())
+	}
+}
+
+type System struct {
+	// you can add some data type if you like
+}
+
+func (System) String() string {
+	return "There's nothing here."
+}
+
+func (System) LoadPTT(url string) PTTArticles {
+	var articles PTTArticles
+	jsonBlob, _ := ioutil.ReadFile(url)
+	checkErr(json.Unmarshal(jsonBlob, &articles))
+	return articles
+}
+
+func (System) LoadFB(url string) FBArticles {
+	var articles FBArticles
+	jsonBlob, _ := ioutil.ReadFile(url)
+	checkErr(json.Unmarshal(jsonBlob, &articles))
+	return articles
+}
+
+/* [TODO]
+func (System) CountCyberWarriors() {
+
+}
+
+func (System) CountKeyWord(){
+
+}
+*/
