+++ lab9/andytsai2000/lab9.go
@@ -1,0 +1,20 @@
+package main
+
+import (
+	"fmt"
+
+	// [TODO] set your module name, go mod init my_module->MOUDULE_NAME
+	"<MOUDULE_NAME>/cw"
+)
+
+func main() {
+	cwSystem := cw.System{}
+
+	PTTArticles := cwSystem.LoadPTT("./data/ptt.json")
+	fmt.Printf("%+v\n", PTTArticles.Articles[0])
+
+	FBArticles := cwSystem.LoadFB("./data/fb.json")
+	fmt.Printf("%+v\n", FBArticles.Articles[0])
+	
+	fmt.Println(cwSystem)
+}
