+++ lab9/andytsai2000/validate.sh
