--- lab9/cw/account.go
@@ -1,12 +1,0 @@
-package cw
-
-type Account struct {
-}
-
-type PTTAccount struct {
-	Account
-}
-
-type FBAccount struct {
-	Account
-}
